Move SelfCycle tests
